Added patching work flow back to the deck.
After writing down my thoughts about the evolution of how we use Git
on open source projects
(http://developerworkflow.com/resources/evolution-social-coding.html)
I realized that the patch work flow was actually interesting and
relevant in the modern context. Although it's 'archaic' by today's
standards, it's actually how the git tools are designed to work. The
social coding tools we use today are NOT optimized for the patching
work flow, and as a result it makes parts of Git more difficult to
use with other permission strategies. Knowing the historical
context, I think, helps to explain why rebasing can be really
important to use.