Lesson 8: Enhancing notes for git stash
The notes were insufficient for this lesson, so I've fixed them. You
need to include the commit hash for an example commit that's bad
(broken) and an example commit that's working (good). If you don't
specify a "bad" commit id, Git will assume that the latest commit for
is "bad". It's better if you can be explicit with both in case there's
something funky going on.